+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Some examples, more below in the actual changelog (newer entries are more likely
 
 ### Added
 * core/v1: add `created_at` and `updated_at` fields to Resource type
+* generic client: make common.PartialResource filterable
 
 ## [0.5.3] - 2023-06-16